parandused

+++ src/main/java/springalgus/Rakendus.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
 @SpringBootApplication
 public class Rakendus{
 	@RequestMapping("/pindala")
-	float korrutamine(int arv1){
+	double korrutamine(int arv1){
 		return (arv1*arv1*3.14);
 	}
     public static void main(String[] args) {
